ue5-relay/requirements.txt
supabase>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
websockets>=12.0
